Load package fonts

Registers fonts contained in penngradlings with systemfonts::register_font(). This function is called once silently when the package is loaded.

Usage

load_pkg_fonts(verbose = TRUE)

Arguments

verbose

Whether the newly registered fonts should be printed to the console. Defaults to TRUE.

Value

An invisible call to systemfonts::registry_fonts() for further inspection.

Details

All styles of a font family are registered as the Regular member of their own families in the form of "family-style(s)".

Call systemfonts::registry_fonts() after loading the fonts for more details (also returned invisibly).

Fonts loaded

Charis SIL

by SIL international, retrieved from https://software.sil.org/charis/.

Inter

by Rasmus Andersson, retrieved from https://fonts.google.com/specimen/Inter.

Piazzolla

by Juan Pablo del Peral, retrieved from https://fonts.google.com/specimen/Piazzolla.

Atkinson Hyperlegible

by Braille Institute, retrieved from https://fonts.google.com/specimen/Atkinson+Hyperlegible.

Bootstrap Icons

by The Bootstrap Authors, retrieved from https://icons.getbootstrap.com/.

Material Icons

by Google Design, retrieved from https://fonts.google.com/icons/.

Font Awesome 5 Free

by Dave Gandy https://fontawesome.com/. The free version contains a subset of the Regular, Solid, and Brand styles: https://fontawesome.com/v5.15/icons?d=gallery&p=2&s=brands,regular,solid&m=free.

Copy of the license for each font is included in the package source. The use of these fonts in the package does not indicate endorsement or liability of the copyright/trademark holders. We thank the developers of the fonts for their open-source contributions and permissive licenses.

Examples

# \dontrun{
load_pkg_fonts()
#> 
#> -- AtkinsonHyperlegible --
#> 
#> v 4 styles registered:
#> * AtkinsonHyperlegible-Bold
#> * AtkinsonHyperlegible-BoldItalic
#> * AtkinsonHyperlegible-Italic
#> * AtkinsonHyperlegible-Regular
#> 
#> -- BootstrapIcons --
#> 
#> v 1 style registered:
#> * BootstrapIcons-Regular
#> 
#> -- CharisSIL --
#> 
#> v 2 styles registered:
#> * CharisSIL-Bold
#> * CharisSIL-Regular
#> 
#> -- FontAwesome5Free --
#> 
#> v 3 styles registered:
#> * FontAwesome5-Brands
#> * FontAwesome5-Regular
#> * FontAwesome5-Solid
#> 
#> -- Inter --
#> 
#> v 18 styles registered:
#> * Inter-Black
#> * Inter-BlackItalic
#> * Inter-Bold
#> * Inter-BoldItalic
#> * Inter-ExtraBold
#> * Inter-ExtraBoldItalic
#> * Inter-ExtraLight
#> * Inter-ExtraLightItalic
#> * Inter-Italic
#> * Inter-Light
#> * Inter-LightItalic
#> * Inter-Medium
#> * Inter-MediumItalic
#> * Inter-Regular
#> * Inter-SemiBold
#> * Inter-SemiBoldItalic
#> * Inter-Thin
#> * Inter-ThinItalic
#> 
#> -- MaterialIcons --
#> 
#> v 2 styles registered:
#> * MaterialIcons-Filled
#> * MaterialIcons-Outlined
#> 
#> -- Piazzolla --
#> 
#> v 12 styles registered:
#> * Piazzolla-Black
#> * Piazzolla-BlackItalic
#> * Piazzolla-Bold
#> * Piazzolla-BoldItalic
#> * Piazzolla-ExtraBold
#> * Piazzolla-ExtraBoldItalic
#> * Piazzolla-Italic
#> * Piazzolla-Medium
#> * Piazzolla-MediumItalic
#> * Piazzolla-Regular
#> * Piazzolla-SemiBold
#> * Piazzolla-SemiBoldItalic
#> 
#> -- PiazzollaSC --
#> 
#> v 6 styles registered:
#> * PiazzollaSC-Black
#> * PiazzollaSC-Bold
#> * PiazzollaSC-ExtraBold
#> * PiazzollaSC-Medium
#> * PiazzollaSC-Regular
#> * PiazzollaSC-SemiBold
#> --------------------------------------------------------------------------------
#> i Done! Check `systemfonts::registry_fonts()` for more details.
# }