Read the results from a PCIbex experiment

Converts a comma-separated PCIbex results file into rectangular format.

read_pcibex(file, exclude_controller_name = TRUE)

Arguments

file
Path to the results file

exclude_controller_name
Whether to skip parsing a column for "Controller name". In the new PCIbex (last checked: September 2021), "Controller name" appears in the column specification as the third column but does not have associated values in the results. This is TRUE by default but when parsing results files from the old PCIbex, it should be set to FALSE.

Value

A dataframe

Examples

# \dontrun{
# It takes a few seconds because it's reading from remote
# but the parsing itself is very fast.
# Also note that this results file is from the old PCIbex
# so `exclude_controller_name` is set to `FALSE`
dplyr::glimpse(
  read_pcibex(
    "https://raw.githubusercontent.com/yjunechoe
     /Semantic-Persistence/master/data/result.txt",
     exclude_controller_name = FALSE
  )
)

#> Rows: 23,302
#> Columns: 17
#> $ Time_results_were_received           <int> 1586129536, 1586129536, 158612953~
#> $ MD5_hash_of_participant_s_IP_address <chr> "ed21346ef6cb189471f1665c27e1065e~
#> $ Controller_name                      <chr> "PennController", "PennController~
#> $ Item_number                          <int> 0, 0, 2, 2, 3, 3, 3, 4, 4, 4, 5, ~
#> $ Element_number                       <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ~
#> $ Type                                 <chr> "consent", "consent", "intro", "i~
#> $ Group                                <chr> "NULL", "NULL", "NULL", "NULL", "~
#> $ PennElementType                      <chr> "PennController", "PennController~
#> $ PennElementName                      <chr> "0", "0", "2", "2", "3", "b7dbb12~
#> $ Parameter                            <chr> "_Trial_", "_Trial_", "_Trial_", ~
#> $ Value                                <chr> "Start", "End", "Start", "End", "~
#> $ EventTime                            <chr> "1586129111186", "1586129125480",~
#> $ Comments                             <chr> "NULL", "NULL", "NULL", "NULL", "~
#> $ Item                                 <chr> NA, NA, NA, NA, NA, NA, NA, NA, N~
#> $ Cond                                 <chr> NA, NA, NA, NA, NA, NA, NA, NA, N~
#> $ Group_1                              <chr> NA, NA, NA, NA, NA, NA, NA, NA, N~
#> $ Participant_ID                       <chr> NA, NA, NA, NA, NA, NA, NA, NA, N~

# }