Construct a list that names itself

Construct a list that names itself

name_self(list_expr, .fn = function(x, y) {
    x
})

Arguments

list_expr
A expression using a call to list, c, or the like.

.fn
A function passed to purrr::imap_chr, called on the character vector returned after coercing the element symbols in list_expr to name. Defaults to the identity function.

Value

A named list

Examples

name_self(c(min, mean, max))

#> $min
#> function (..., na.rm = FALSE)  .Primitive("min")
#> 
#> $mean
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x7fe729b2b158>
#> <environment: namespace:base>
#> 
#> $max
#> function (..., na.rm = FALSE)  .Primitive("max")
#> 

name_self(c(min, mean, max), .fn = ~ toupper(.x))

#> $MIN
#> function (..., na.rm = FALSE)  .Primitive("min")
#> 
#> $MEAN
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x7fe729b2b158>
#> <environment: namespace:base>
#> 
#> $MAX
#> function (..., na.rm = FALSE)  .Primitive("max")
#> 

name_self(c(min, mean, max), .fn = function(x, y) { paste0(x, y) })

#> $min1
#> function (..., na.rm = FALSE)  .Primitive("min")
#> 
#> $mean2
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x7fe729b2b158>
#> <environment: namespace:base>
#> 
#> $max3
#> function (..., na.rm = FALSE)  .Primitive("max")
#> 

# Helps when naming produced by from `across()`
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

## Without `name_self()`
mtcars %>%
  group_by(cyl) %>%
  summarize(across(where(~ max(.x) > 100), list(min, mean, max)))

#> # A tibble: 3 × 7
#>     cyl disp_1 disp_2 disp_3  hp_1  hp_2  hp_3
#>   <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1     4   71.1   105.   147.    52  82.6   113
#> 2     6  145     183.   258    105 122.    175
#> 3     8  276.    353.   472    150 209.    335

## With `name_self()`
mtcars %>%
  group_by(cyl) %>%
  summarize(across(where(~ max(.x) > 100), name_self(list(min, mean, max))))

#> # A tibble: 3 × 7
#>     cyl disp_min disp_mean disp_max hp_min hp_mean hp_max
#>   <dbl>    <dbl>     <dbl>    <dbl>  <dbl>   <dbl>  <dbl>
#> 1     4     71.1      105.     147.     52    82.6    113
#> 2     6    145        183.     258     105   122.     175
#> 3     8    276.       353.     472     150   209.     335